+++ NEWS.md
@@ -1,5 +1,6 @@
 # qlcData 0.2
 
+* adding function pass_align() to transfer alignment to a string
 * changed vignette naming
 
 # qlcData 0.1